feat: added abac permissions for ui order actions (SL-6132)

+++ modules/order/controllers/OrderController.php
@@ -5,6 +5,8 @@ namespace modules\order\controllers;
 use common\models\Lead;
 use modules\fileStorage\src\entity\fileOrder\FileOrder;
 use modules\fileStorage\src\entity\fileStorage\FileStorageQuery;
+use modules\order\src\abac\dto\OrderAbacDto;
+use modules\order\src\abac\OrderAbacObject;
 use modules\order\src\entities\order\OrderSourceType;
 use modules\order\src\entities\order\OrderStatus;
 use modules\order\src\entities\order\search\OrderCrudSearch;
@@ -24,6 +26,7 @@ use yii\db\Exception;
 use yii\helpers\ArrayHelper;
 use yii\helpers\VarDumper;
 use yii\web\BadRequestHttpException;
+use yii\web\ForbiddenHttpException;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 use yii\web\Response;
@@ -60,6 +63,13 @@ class OrderController extends FController
                     'delete-ajax' => ['POST'],
                 ],
             ],
+            'access' => [
+                'allowActions' => [
+                    'view',
+                    'update-ajax',
+                    'delete'
+                ]
+            ]
         ];
         return ArrayHelper::merge(parent::behaviors(), $behaviors);
     }
@@ -140,6 +150,11 @@ class OrderController extends FController
             return $throwable->getMessage();
         }
 
+        $orderAbacDto = new OrderAbacDto($modelOrder);
+        if (!Yii::$app->abac->can($orderAbacDto, OrderAbacObject::ACT_UPDATE, OrderAbacObject::ACTION_ACCESS)) {
+            throw new ForbiddenHttpException('Access denied');
+        }
+
         $model = new OrderForm();
         $model->or_lead_id = $modelOrder->or_lead_id;
         $model->or_id = $modelOrder->or_id;
@@ -207,6 +222,10 @@ class OrderController extends FController
 
         try {
             $model = $this->findModel($id);
+            $orderAbacDto = new OrderAbacDto($model);
+            if (!Yii::$app->abac->can($orderAbacDto, OrderAbacObject::ACT_DELETE, OrderAbacObject::ACTION_ACCESS)) {
+                throw new ForbiddenHttpException('Access denied');
+            }
             if (!$model->delete()) {
                 throw new Exception('Order (' . $id . ') not deleted', 2);
             }
@@ -276,6 +295,11 @@ class OrderController extends FController
     {
         if (!$order = Order::findOne(['or_gid' => $gid])) {
             throw new NotFoundHttpException('Order not found by GID(' . $gid . ')');
+        }
+
+        $orderAbacDto = new OrderAbacDto($order);
+        if (!Yii::$app->abac->can($orderAbacDto, OrderAbacObject::ACT_DETAIL_VIEW, OrderAbacObject::ACTION_ACCESS)) {
+            throw new ForbiddenHttpException('Access denied');
         }
 
         return $this->render('view', [
